Address issue #9 -- external ref clk
This commit adds an LMK clock configuration file which locks
to an external 10MHz signal provided on an RFSoC4x2's CLK_IN connector.

Add this file to the RFSoC filesystem under `/lib/firmware/`, and point
to it in a configuration yaml using the `lmkfile` key.

Tested on a scope by verifying that a 30MHz DAC tone is phase locked to the 10MHz ref